CLN : literalinclude handler_override
Also move the configure_traits_view_buttons file - which i missed before

	renamed:    examples/tutorials/doc_examples/examples/configure_traits_view_buttons.py -> docs/source/traitsui_user_manual/examples/configure_traits_view_buttons.py
	new file:   docs/source/traitsui_user_manual/examples/handler_override.py
	modified:   docs/source/traitsui_user_manual/handler.rst
	deleted:    examples/tutorials/doc_examples/examples/handler_override.py
