+++ docs/source/traitsui_user_manual/examples/configure_traits_view_buttons.py
