+++ docs/source/traitsui_user_manual/handler.rst
@@ -442,41 +442,8 @@ overridden setattr() method and user interface notification method.
 
 .. rubric:: Example 9: Using a Handler that reacts to trait changes
 
-::
-
-    # handler_override.py -- Example of a Handler that overrides
-    # setattr(), and that has a user interface notification method
-
-    from traits.api import Bool, HasTraits
-    from traitsui.api import Handler, View
-
-    class TC_Handler(Handler):
-
-        def setattr(self, info, object, name, value):
-            Handler.setattr(self, info, object, name, value)
-            info.object._updated = True
-
-        def object__updated_changed(self, info):
-            if info.initialized:
-                info.ui.title += "*"
-
-    class TestClass(HasTraits):
-        b1 = Bool()
-        b2 = Bool()
-        b3 = Bool()
-        _updated = Bool(False)
-
-    view1 = View(
-        'b1',
-        'b2',
-        'b3',
-        title="Alter Title",
-        handler=TC_Handler(),
-        buttons=['OK', 'Cancel'],
-    )
-
-    tc = TestClass()
-    tc.configure_traits(view=view1)
+.. literalinclude:: examples/handler_override.py
+   :start-at: handler_override.py
 
 .. image:: images/alter_title_before.png
    :alt: Dialog box with empty checkboxes and a title of "Alter Title"
